Move util.read_private_key() to crypto
Move the read_private_key() function from util to crypto.  This further
encapsulates access to the cryptography library, and prepares the CLI to
be able to function without it.
