+++ brkt_cli/get_public_key/__init__.py
@@ -62,7 +62,7 @@ class GetPublicKeySubcommand(Subcommand):
         return values.make_private_key_verbose
 
     def run(self, values):
-        crypto = util.read_private_key(values.private_key_path)
+        crypto = brkt_cli.crypto.read_private_key(values.private_key_path)
         util.write_to_file_or_stdout(crypto.public_key_pem, values.out)
         return 0
