+++ brkt_cli/test_util.py
@@ -11,12 +11,10 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the
 # License for the specific language governing permissions and
 # limitations under the License.
-import tempfile
 import time
 import unittest
 
 from brkt_cli import util
-from brkt_cli.crypto import test_crypto
 from brkt_cli.validation import ValidationError
 
 
@@ -146,58 +144,3 @@ class TestRetry(unittest.TestCase):
         """ Test retry based on the exception type. """
         self._get_wrapped(on=[TestException])(5)
         self.assertEqual(6, self.num_calls)
-
-
-# This PEM was generated by openssl, and must only be used for
-# unit testing.
-TEST_SECT163K1_PRIVATE_KEY_PEM = """-----BEGIN EC PRIVATE KEY-----
-MFMCAQEEFQGZdbOd9KrGmynCqcy3S1QDvq67xqAHBgUrgQQAAaEuAywABAej5bYZ
-ZN4/1PqYAE8Cz4aSMC8KjgCDfMkzv0VURQGcSQ5xc4aCnI7bIQ==
------END EC PRIVATE KEY-----
-"""
-
-
-class TestReadPrivateKey(unittest.TestCase):
-
-    def test_read_private_key(self):
-        """ Test reading the signing key from a file. """
-        # Write private key to a temp file.
-        key_file = tempfile.NamedTemporaryFile()
-        key_file.write(test_crypto.TEST_PRIVATE_KEY_PEM)
-        key_file.flush()
-
-        crypto = util.read_private_key(key_file.name)
-        self.assertEqual(
-            test_crypto.TEST_PRIVATE_KEY_X, crypto.x)
-        key_file.close()
-
-    def test_read_private_key_invalid_curve(self):
-        """ Test that we require NIST384p for the signing key. """
-        # Write private key to a temp file.
-        key_file = tempfile.NamedTemporaryFile()
-        key_file.write(TEST_SECT163K1_PRIVATE_KEY_PEM)
-        key_file.flush()
-
-        with self.assertRaises(ValidationError):
-            util.read_private_key(key_file.name)
-        key_file.close()
-
-    def test_read_private_key_io_error(self):
-        """ Test that we handle IOError when reading the signing key.
-        """
-        # Read from a directory.
-        with self.assertRaises(ValidationError):
-            util.read_private_key('.')
-
-        # Read from a file that doesn't exist.
-        with self.assertRaises(ValidationError):
-            util.read_private_key('nothing_here.pem')
-
-        # Read from a malformed file.
-        key_file = tempfile.NamedTemporaryFile()
-        key_file.write('abc')
-        key_file.flush()
-
-        with self.assertRaises(ValidationError):
-            util.read_private_key(key_file.name)
-        key_file.close()
